1st draft of code
As a Data science graduate, checking the data for inaccuracies in a must. A small search was done to check for basic types of errors in data which are: null values, duplicates and data outside the logical range. Afterwards, a small EDA was conducted to understand the data on a deeper level. What I concluded: 
-Data is relatively clean
-Distribution of ratings is not equal between games
-Some users tend to give reviews/ratings more often than others
-Average Rating is not a good standard for ranking

A basic form of class was introduced to store each game's data. Until now,game id, name, average ranking and wilson ranking is all what i see needed to describe each game. Method are still now implemented. 

For future updates to complete the project, what i need is: two functions to calculate each rank, a certain method to put these values in the class, and another to retrieve them. Since the class is not sorted, i need a way to retrieve the ranking of all games in a sorted way so i can see the highest ranks. For requirements purposes, I have to visualize the differences of ranking in some plots.